% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_equal_weight_group.R
\name{add_equal_weight_group}
\alias{add_equal_weight_group}
\title{add_equal_weight_group()}
\usage{
add_equal_weight_group(table, sort_by, expo = NULL,
  group_variable_name = "groupe", nb = 10)
}
\arguments{
\item{table}{data.frame  source}

\item{sort_by}{Variable utilisée pour trier les observations.}

\item{expo}{Exposition (utilisée pour créer des quantiles de la même taille.  Si NULL, l'exposition est égale pour toutes les observations) (Défault = NULL).}

\item{group_variable_name}{Nom de la variable de groupes créée}

\item{nb}{Nombre de quantiles crées (défaut = 10)}
}
\description{
Cette fonction crée des groupe (quantiles) avec le nombre nombre total d'exposition.
}
