% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_cutoff.R
\name{df_cutoff}
\alias{df_cutoff}
\title{df_cutoff()}
\usage{
df_cutoff(y, label = "label")
}
\description{
Cette fonction crée un dataframe avec 2 observations pour créer une ligne horizontale dans un ggplot
}
\examples{
+geom_line(data= df_cutoff(100), aes(x=x, y=y), linetype = 3, color = "black")
}
