% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_line45.R
\name{df_line45}
\alias{df_line45}
\title{df_line45()}
\usage{
df_line45(label = "label")
}
\description{
Cette fonction crée un dataframe avec 2 observations pour créer une ligne à 45 degrés dans un ggplot
}
\examples{
+geom_line(data = df_line45(),aes(x=x, y=y), color = "lightgray")
}
