% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disloc.R
\name{disloc}
\alias{disloc}
\title{disloc()}
\usage{
disloc(data, pred1, pred2, expo, obs, nb = 10, pred1_lab = "pred1",
  pred2_lab = "pred2", y_label = "sinistralité",
  y_format = scales::percent_format, graphe_title = NULL)
}
\arguments{
\item{data}{data.frame  source}

\item{pred1}{prediction of first model}

\item{expo}{exposure var}

\item{obs}{observed result}

\item{nb}{nombre de quantils créés}

\item{pred1_lab}{Label pour la première prédiction dans le graphique}

\item{pred2_lab}{Label pour la deuxième prédiction dans le graphique}

\item{y_label}{Label pour la valeur réalisée dans le graphique}

\item{y_format}{Fonction utilisée pour formater l'axe des y dans le graphique (par exemple percent_format() ou dollar_format() du package scales)}

\item{graphe_title}{une string qui donne le titre du modèle}

\item{pred1}{prediction of second model}
}
\description{
Cette fonction crée un tableau et une double lift chart
}
