% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disloc_buckets.R
\name{disloc_buckets}
\alias{disloc_buckets}
\title{disloc_buckets()}
\usage{
disloc_buckets(data, pred1, pred2, expo, obs, buckets = c(0, 0.825,
  0.875, 0.925, 0.975, 1.025, 1.075, 1.125, 1.175, Inf),
  pred1_lab = "pred1", pred2_lab = "pred2", actual_lab = "réalisé",
  x_label = "ratio entre les prédictions", y_label = "sinistralité",
  y_format = percent_format(), graphe_title = NULL)
}
\arguments{
\item{data}{data.frame  source}

\item{pred1}{prediction of first model}

\item{expo}{exposure}

\item{obs}{observed result}

\item{buckets}{limites des buckets}

\item{pred1_lab}{Label pour la première prédiction dans le graphique}

\item{pred2_lab}{Label pour la deuxième prédiction dans le graphique}

\item{x_label}{Label pour l'axe des x}

\item{y_label}{Label pour la valeur réalisée dans le graphique}

\item{y_format}{Fonction utilisée pour formater l'axe des y dans le graphique (par exemple percent_format() ou dollar_format() du package scales)}

\item{graphe_title}{une string qui donne le titre du modèle}

\item{pred1}{prediction of second model}
}
\description{
Cette fonction crée un tableau et une double lift chart en séparant sur des buckets de ratio pré-définis au lieu d'une exposition constante.
}
