% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_double_lift_chart_data.R
\name{get_double_lift_chart_data}
\alias{get_double_lift_chart_data}
\title{get_double_lift_chart_data()}
\usage{
get_double_lift_chart_data(table, expo = NULL, nb = 10,
  pred1 = "pred1", pred2 = "pred2", obs = "obs")
}
\arguments{
\item{table}{data.frame à partir duquel on construit le lift}

\item{expo}{Exposition (utilisée pour créer des quantiles de la même taille.  Si NULL, l'exposition est égale pour toutes les observations) (Défault = NULL).}

\item{nb}{Nombre de quantiles crées (défaut = 10)}

\item{pred1}{Prédiction du premier modèle.}

\item{pred2}{Prédiction du second modèle .}

\item{obs}{Variable réponse observée (défaut = gperts)}
}
\description{
Cette fonction retourne un data.frame avec les colonnes nécessaires pour créer un beau double lift chart.
}
\keyword{chart}
\keyword{double}
\keyword{lift}
