% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gam_fold_pred_model.R
\name{get_gam_fold_pred_model}
\alias{get_gam_fold_pred_model}
\title{get_gam_fold_pred_model()}
\usage{
get_gam_fold_pred_model(data, nb_fold = 10, formula)
}
\arguments{
\item{data}{la base de données (categorical variables have been replaced by dummies.)}

\item{nb_fold}{le nombre de fold dans notre validation croisée}

\item{formula}{la formule, probablement générée par get_gam_formula()}
}
\description{
get_gam_fold_pred_model() retourne les prédictions et les modèles de chaque fold d'un gam.  Pourquoi utiliser ça plutôt que caret?  Je pense que c'est parce que caret ne fait pas de gam.. ?  En tout cas, mon modèle calcule présentement des poissons.
}
