% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gam_formula.R
\name{get_gam_formula}
\alias{get_gam_formula}
\title{get_gam_formula()}
\usage{
get_gam_formula(label_var, offset_var, feature_vars)
}
\arguments{
\item{label_var}{le nom de la variable à prédire}

\item{offset_var}{la variable dont le log sera mis en offset}

\item{feature_vars}{un vecteur de noms de variables pour lesquelles une spline sera créée}
}
\description{
get_gam_formula() crée une formule à passer à un gam.  Les feature_vars seront toutes splinées et le log de la offset_Var sera mis en offset.
}
