% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lift_chart_data.R
\name{get_lift_chart_data}
\alias{get_lift_chart_data}
\title{get_lift_chart_data()}
\usage{
get_lift_chart_data(table, sort_by = "rate", expo = NULL, nb = 10,
  pred = "pred", obs = "obs")
}
\arguments{
\item{table}{data.frame à partir duquel on construit le lift}

\item{sort_by}{Variable utilisée pour trier les observations.  utiliser un taux pour lifter contre "aucun modèle" et utiliser un ratio de prédictions pour lifter contre un autre modèle}

\item{expo}{Exposition (utilisée pour créer des quantiles de la même taille.  Si NULL, l'exposition est égale pour toutes les observations) (Défault = NULL).}

\item{nb}{Nombre de quantiles crées (défaut = 10)}

\item{pred}{Prédiction.  Utiliser la prédiction du modèle *de référence* pour une comparaison de modèles.  Utiliser la prédiction du modèle pour comparer contre "no model".}

\item{obs}{Variable réponse observée (défaut = gperts)}
}
\description{
Cette fonction retourne un data.frame avec les colonnes nécessaires pour créer un beau lift chart.  On peut lifter à partir d'un autre modèle ou bien à partir de "no model".
}
\keyword{chart}
\keyword{lift}
