% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partial_dependence_xgb.R
\name{get_partial_dependence_xgb}
\alias{get_partial_dependence_xgb}
\title{get_partial_dependence_xgb()}
\usage{
get_partial_dependence_xgb(xgb_model, pdp_vars = NULL, feature_vars,
  offset_var = NULL, model_data, nquantiles = 21)
}
\arguments{
\item{xgb_model}{le modèle xgboost qui va être prédit}

\item{pdp_vars}{un vecteur caractère avec les noms des variables pour lesquelles on souhaite calculer la pdp}

\item{feature_vars}{un vecteur caractère avec les noms des variables prédictives du modèle}

\item{offset_var}{variable à mettre en offset, si nécessaire}

\item{model_data}{un data frame avec les observations à prédire}

\item{nquantiles}{le nombre de quantiles à calculer.}
}
\description{
Cette fonction calcule les dépendances partielles d'un modèles xgboost après avoir enlevé les outliers. défaut = toutes les variables.   elle retourne un data frame avec les valeurs (Mod) de la variable explicative (Variable)
}
\examples{
zzzz1 <- get_partial_dependence_xgb(xgb_model = resBoost,
                                    feature_vars = c("brake_intensity", "number_of_brake", "number_of_accel", "distraction"),
                                    model_data = model_data,
                                    pdp_vars = c("brake_intensity"),
                                    offset_var = "pred_saison",
                                    nquantiles = 11)
}
