% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partial_dependence_xgb_2d.R
\name{get_partial_dependence_xgb_2d}
\alias{get_partial_dependence_xgb_2d}
\title{get_partial_dependence_xgb_2d()}
\usage{
get_partial_dependence_xgb_2d(xgb_model, feature_vars, model_data,
  pdp_var1, pdp_var2, offset_var = NULL, nquantiles = 5)
}
\arguments{
\item{xgb_model}{le modèle xgboost qui va être prédit}

\item{feature_vars}{un vecteur caractère avec les noms des variables prédictives du modèle}

\item{model_data}{un data frame avec les observations à prédire}

\item{pdp_var1}{la premiere variable pour la pdp (caractère)}

\item{pdp_var2}{la deuxième variable pour la pdp (caractère)}

\item{offset_var}{variable à mettre en offset, si nécessaire}

\item{nquantiles}{le nombre de quantiles à calculer.}
}
\description{
Cette fonction calcule les dépendances partielles en 2D pour 2 variables d'un modèles xgboost après avoir enlevé les outliers.
}
\examples{
zzzz3 <- get_partial_dependence_xgb_2d(xgb_model = resBoost,
                                      feature_vars = c("brake_intensity", "number_of_brake", "number_of_accel", "distraction"),
                                      model_data = model_data,
                                      pdp_var1 = "braking_combined_roadtype0",
                                      pdp_var2 = "brake_intensity",
                                      offset_var = "pred_saison",
                                      nquantiles = 5)
}
