% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xgboost_fold_pred_model.R
\name{get_xgboost_fold_pred_model}
\alias{get_xgboost_fold_pred_model}
\title{get_xgboost_fold_pred_model()}
\usage{
get_xgboost_fold_pred_model(params = list(), data, label_var,
  feature_vars, offset_var = NULL, nb_fold = 10,
  fold_group_var = NULL, seed = 8484, nround = 1000,
  contraintes = NULL)
}
\arguments{
\item{params}{les paramètre du xgboost,}

\item{data}{la base de données (categorical variables have been replaced by dummies.)}

\item{label_var}{le nom de la variable à prédire}

\item{feature_vars}{le vecteur des noms des variables explicatives}

\item{offset_var}{le nom de la variable qui sera mise dans le base_margin (ex:exposure), default= NULL}

\item{nb_fold}{le nombre de fold dans notre validation croisée}

\item{seed}{le numéro de seed}

\item{nround}{le nombre d'arbres à créer avec le xgboost}

\item{contraintes}{vecteur des contraints de monotonicité}
}
\description{
get_xgboost_fold_pred_model() retourne les prédictions et les modèles de chaque fold d'un xgboost.  Pourquoi utiliser ça plutôt que caret?  Parce que je gère le offset et les contraintes de monotonicité.
}
