% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_latest.R
\name{load_latest}
\alias{load_latest}
\title{save_with_date()}
\usage{
load_latest(object, path)
}
\arguments{
\item{object}{le nom de l'objet sauvegardé}

\item{path}{le path où a été sauvegardé l'object  (par rapport à here::here)}
}
\description{
Cette fonction crée charge un object sauvegardé avec le pattern de save_with_date()
}
\examples{
expo_claim_count <- load_latest(expo_claim_count, "Result/ON")
}
