% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_compute_gam_model.R
\name{my_compute_gam_model}
\alias{my_compute_gam_model}
\title{my_compute_gam_model()}
\usage{
my_compute_gam_model(data, label_var, offset_var, feature_vars,
  nb_fold = 10, get_fold_pred = TRUE)
}
\description{
my_compute_gam_model()
}
