% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_compute_xgboost_model.R
\name{my_compute_xgboost_model}
\alias{my_compute_xgboost_model}
\title{my_compute_xgboost_model()}
\usage{
my_compute_xgboost_model(params = list(), data, label_var, feature_vars,
  offset_var = NULL, nb_fold = 10, contraintes = NULL, seed = 8484,
  max_iter = 2000, forced_max_iter = FALSE, get_fold_pred = TRUE)
}
\description{
my_compute_xgboost_model()
}
