% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgsimon-package.R
\docType{package}
\name{pkgsimon-package}
\alias{pkgsimon}
\alias{pkgsimon-package}
\title{pkgsimon: Les fonctions de Simon}
\description{
This packages contains functions that I frequently reuse.
}
\author{
\strong{Maintainer}: Simon Coulombe \email{simoncoulombe@gmail.com}

}
\keyword{internal}
