% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_align_grid.R
\name{plot_align_grid}
\alias{plot_align_grid}
\title{plot_align_grid()}
\usage{
plot_align_grid(..., heights = NULL)
}
\description{
Cette fonction prend n graphiques créés avec ggplot et les imprime un en dessous de l'autre, en alignant l'axe des x.
}
\examples{
plot_align_grid(plot1  + theme(axis.ticks.x = element_blank(), axis.text.x = element_blank(), axis.title.x = element_blank()) , plot2, heights= c(3,1))
}
