% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove_outliers()}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{le vecteur de valeurs}

\item{na.rm}{est-ce qu'on enlève les na dans le calcul des quantiles?}
}
\description{
Cette fonction prend un vecteur x en entrée et en enlève les outliers ( x > Q3+ 1.5 IQR ou x < Q1 - 1.5IQR)
}
