% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_with_date.R
\name{save_with_date}
\alias{save_with_date}
\title{save_with_date()}
\usage{
save_with_date(object, path)
}
\arguments{
\item{object}{le nom de l'obect à sauvegarder}

\item{path}{le path où sauvegarder l'objet (par rapport à here::here)}
}
\description{
Cette fonction crée sauvegarde un objet avec un prefixe correspondant à la date. voir aussi load_latest()
}
\examples{
save_with_date(expo_claim_count, "Result/ON")
}
