% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dviz.R
\name{theme_dviz}
\alias{theme_dviz}
\title{theme_dviz()}
\usage{
theme_dviz(font_size = 14, font_family = "")
}
\description{
ggplot theme taken from https://github.com/clauswilke/dataviz/blob/master/R/themes.R
}
