% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dviz_grid.R
\name{theme_dviz_grid}
\alias{theme_dviz_grid}
\title{theme_dviz_grid()}
\usage{
theme_dviz_grid(font_size = 14, font_family = "")
}
\description{
grid lines along major axis ticks, no axes - ggplot theme taken from https://github.com/clauswilke/dataviz/blob/master/R/themes.R
}
