% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dviz_half_open.R
\name{theme_dviz_half_open}
\alias{theme_dviz_half_open}
\title{theme_dviz_half_open()}
\usage{
theme_dviz_half_open(font_size = 14, font_family = "")
}
\description{
ggplot theme taken from https://github.com/clauswilke/dataviz/blob/master/R/themes.R
}
