% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dviz_hgrid.R
\name{theme_dviz_hgrid}
\alias{theme_dviz_hgrid}
\title{theme_dviz_hgrid()}
\usage{
theme_dviz_hgrid(font_size = 14, font_family = "")
}
\description{
Horizontal lines only - ggplot theme taken from https://github.com/clauswilke/dataviz/blob/master/R/themes.R
}
