% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dviz_vgrid.R
\name{theme_dviz_vgrid}
\alias{theme_dviz_vgrid}
\title{theme_dviz_vgrid()}
\usage{
theme_dviz_vgrid(font_size = 14, font_family = "")
}
\description{
vertical lines only - ggplot theme taken from https://github.com/clauswilke/dataviz/blob/master/R/themes.R
}
