% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\title{wrapper()}
\usage{
wrapper(x, ...)
}
\description{
Cette fonction ajoute des line breaks pour avoir des titre qui fittent bien dans les ggplot..
}
\examples{
wrapper("pouet pouet", 50))
}
